refactoring: abstract TaskManagerTest
